requirements.txt
aiofiles==23.2.1
aiogram==3.8.0
aiohttp==3.9.3
aiosignal==1.3.1
aiosqlite==0.17.0
annotated-types==0.6.0
async-timeout==4.0.3
attrs==23.2.0
betterlogging==1.0.0
cachetools==5.3.3
certifi==2024.2.2
click==8.1.7
colorama==0.4.6
frozenlist==1.4.1
idna==3.6
iso8601==1.1.0
magic-filter==1.0.12
multidict==6.0.5
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.7.4
pydantic-settings==2.3.4
pydantic_core==2.18.4
pypika-tortoise==0.1.6
python-dotenv==1.0.1
pytz==2024.1
redis==5.0.7
tomli==2.0.1
tortoise-orm==0.21.4
typing_extensions==4.10.0
yarl==1.9.4

APScheduler~=3.10.4
tzlocal~=5.2